Modularize store
