+++ src/store/authentication/getters.js
@@ -1,0 +1,3 @@
+export const authStatus = state => {
+  state.status
+}
