+++ src/store/authentication/state.js
@@ -1,0 +1,4 @@
+export default {
+  token: sessionStorage.getItem('token') || '',
+  status: ''
+};
