+++ src/store/automations/mutations.js
@@ -1,0 +1,3 @@
+export const setAutomationsList = (state, {list}) => {
+  state.automationsList = list
+}
