+++ src/store/commands/mutations.js
@@ -1,0 +1,3 @@
+export const setCommandsList = (state, {list}) => {
+  state.commandsList = list
+}
