+++ src/store/commands/state.js
@@ -1,0 +1,3 @@
+export default {
+  commandsList: ''
+};
