+++ src/store/dashboard/actions.js
@@ -1,0 +1,45 @@
+export const openAutomations = (context) => {
+  context.commit('closeSection')
+  context.commit('showAutomationsPanel')
+}
+
+export const openLanguages = (context) => {
+  context.commit('closeSection')
+  context.commit('showLanguagesPanel')
+}
+
+export const openPlugins = (context) => {
+  context.commit('closeSection')
+  context.commit('showPluginsPanel')
+}
+
+export const openSettings = (context) => {
+  context.commit('closeSection')
+}
+
+export const openSystem = (context) => {
+  context.commit('closeSection')
+  context.commit('showSystemPanel')
+}
+
+export const openRoles = (context) => {
+  context.commit('closeSection')
+  context.commit('showRolesPanel')
+}
+
+export const openUsers = (context) => {
+  context.commit('closeSection')
+  context.commit('showUsersPanel')
+}
+
+export const shutdownFD = (context) => {
+  console.log('Implement this')
+}
+
+export const setAdvancedMode = (context) => {
+  context.commit('setAdvancedMode')
+}
+
+export const setMobileSection = (context, section) => {
+  context.commit('setMobileSection', section)
+}
