+++ src/store/dashboard/mutations.js
@@ -1,5 +1,3 @@
-import Vue from 'vue'
-
 export const hideAll = state => {
   hideAllDashboardDisplays(state)
 }
@@ -96,101 +94,3 @@ function hideAllDashboardDisplays (state) {
   state.rolesOpen = false
   state.usersOpen = false
 }
-
-// Mutations for API
-
-// ENVIRONMENTS
-export const setEnvironmentsList = (state, {list}) => {
-  state.environmentsList = list
-}
-
-export const setEnvironmentRoomsList = (state, {list}) => {
-  state.roomsList = list
-}
-// END ENVIRONMENTS
-
-export const setCommandsList = (state, {list}) => {
-  state.commandsList = list
-}
-
-export const setTriggersList = (state, {list}) => {
-  state.triggersList = list
-}
-
-export const setAutomationsList = (state, {list}) => {
-  state.automationsList = list
-}
-
-// PLUGINS
-export const setPluginsList = (state, {list}) => {
-  state.pluginsList = list
-}
-
-export const updatePluginFromWS = (state, plugin) => {
-  var idx
-  idx = state.pluginsList.findIndex(item => item.pluginName === plugin.pluginName)
-  // state.pluginsList[idx] = plugin
-  Vue.set(state.pluginsList, idx, plugin)
-}
-
-export const setMarketplaceCategoriesList = (state, {list}) => {
-  state.marketplaceCategoriesList = list
-}
-
-export const setMarketplaceCategoryPluginsList = (state, {list}) => {
-  state.marketplaceCategoryPluginsList = list
-}
-// END PLUGINS
-
-// THINGS
-export const setThingsList = (state, {list}) => {
-  state.thingsList = list
-}
-
-export const updateThingFromWS = (state, thing) => {
-  var idx
-  idx = state.thingList.findIndex(item => item.name === thing.name)
-  // state.thingsList[idx] = thing
-  Vue.set(state.thingsList, idx, thing)
-}
-
-export const setThingTemplatesList = (state, {list}) => {
-  state.thingTemplatesList = list
-}
-// END THINGS
-
-export const setUsersList = (state, {list}) => {
-  state.usersList = list
-}
-
-export const setRolesList = (state, {list}) => {
-  state.rolesList = list
-}
-
-export const setSystemInfo = (state, data) => {
-  state.systemInfo = data
-}
-
-export const setEnvironmentThingsList = (state, {list}) => {
-  state.environmentThingsList = list
-}
-// end of mutations for API
-
-// AUTHENTICATION
-export const authRequest = (state) => {
-  state.status = 'loading'
-}
-
-export const authError = (state) => {
-  state.status = 'error'
-}
-
-export const authSuccess = (state, token) => {
-  state.status = 'success'
-  state.token = token
-}
-
-export const authLogout = (state) => {
-  state.token = ''
-}
-// END AUTHENTICATION
