+++ src/store/dashboard/state.js
@@ -1,0 +1,22 @@
+export default {
+  // Dashboard state
+  isMobile: '', // REMOVE ME
+  mobileSection: 'home',
+  displaySettings: false,
+  displayInfo: false,
+  displayAlerts: false,
+  displayLogout: false,
+  displayThings: false,
+  blur: false,
+  // end of Dashboard state
+  language: 'en',
+  // Settings state
+  systemOpen: false,
+  automationsOpen: false,
+  languagesOpen: false,
+  pluginsOpen: false,
+  rolesOpen: false,
+  usersOpen: false,
+  // end of Settings state
+  advancedMode: false
+};
