+++ src/store/environments/mutations.js
@@ -1,0 +1,11 @@
+export const setEnvironmentsList = (state, {list}) => {
+  state.environmentsList = list
+}
+
+export const setEnvironmentRoomsList = (state, {list}) => {
+  state.roomsList = list
+}
+
+export const setEnvironmentThingsList = (state, {list}) => {
+  state.environmentThingsList = list
+}
