+++ src/store/environments/state.js
@@ -1,0 +1,4 @@
+export default {
+  environmentsList: '',
+  roomsList: ''
+};
