+++ src/store/plugins/state.js
@@ -1,0 +1,5 @@
+export default {
+  marketplaceCategoriesList: '',
+  marketplaceCategoryPluginsList: '',
+  pluginsList: []
+};
