+++ src/store/resources/actions.js
@@ -1,0 +1,5 @@
+import axios from '@/utils/fd-axios'
+
+export const getResource = (context, resourceId) => {
+  return axios.defaults.baseURL + '/resources/' + resourceId
+}
