+++ src/store/resources/state.js
@@ -1,0 +1,1 @@
+export default {};
