+++ src/store/roles/mutations.js
@@ -1,0 +1,3 @@
+export const setRolesList = (state, {list}) => {
+  state.rolesList = list
+}
