+++ src/store/roles/state.js
@@ -1,0 +1,3 @@
+export default {
+  rolesList: ''
+};
