+++ src/store/system/mutations.js
@@ -1,0 +1,3 @@
+export const setSystemInfo = (state, data) => {
+  state.systemInfo = data
+}
