+++ src/store/things/state.js
@@ -1,0 +1,4 @@
+export default {
+  thingsList: [],
+  thingTemplatesList: ''
+};
