+++ src/store/triggers/getters.js
@@ -1,0 +1,7 @@
+export const getHardwareTriggers = state => {
+  return state.triggersList.filter(trigger => {
+    if (trigger.hardwareLevel === true) {
+      return trigger
+    }
+  })
+}
