+++ src/store/triggers/mutations.js
@@ -1,0 +1,3 @@
+export const setTriggersList = (state, {list}) => {
+  state.triggersList = list
+}
