+++ src/store/users/index.js
@@ -1,0 +1,11 @@
+import state from './state';
+import * as getters from './getters';
+import * as mutations from './mutations';
+import * as actions from './actions';
+
+export default {
+  state,
+  getters,
+  mutations,
+  actions,
+};
