+++ src/store/users/mutations.js
@@ -1,0 +1,3 @@
+export const setUsersList = (state, {list}) => {
+  state.usersList = list
+}
