+++ src/store/users/state.js
@@ -1,0 +1,3 @@
+export default {
+  usersList: ''
+};
